# ShortLink/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Short Link",
    "version": "v1"
  },
  "paths": {
    "/api/v1/auth/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginParam"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginParam"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/LoginParam"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/v1/auth/register": {
      "post": {
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterParam"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterParam"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterParam"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/{hash}": {
      "get": {
        "tags": [
          "Url"
        ],
        "parameters": [
          {
            "name": "hash",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "api-version",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/v1/url/{hash}": {
      "get": {
        "tags": [
          "Url"
        ],
        "parameters": [
          {
            "name": "hash",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/v1/url/shorten": {
      "post": {
        "tags": [
          "Url"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ShortenParam"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ShortenParam"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ShortenParam"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/v1/url/update-metadata": {
      "post": {
        "tags": [
          "Url"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMetaDataParam"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMetaDataParam"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMetaDataParam"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "LoginParam": {
        "required": [
          "loginName",
          "password"
        ],
        "type": "object",
        "properties": {
          "loginName": {
            "minLength": 1,
            "type": "string"
          },
          "password": {
            "minLength": 1,
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "RegisterParam": {
        "required": [
          "loginName",
          "password"
        ],
        "type": "object",
        "properties": {
          "loginName": {
            "minLength": 1,
            "type": "string"
          },
          "password": {
            "minLength": 1,
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "ShortenParam": {
        "type": "object",
        "properties": {
          "url": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "UpdateMetaDataParam": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "thumbnail": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      }
    }
  }
}